Функции для работы с изображение сделаны

+++ General/Service/File/ImageFileService.cs
@@ -13,8 +13,8 @@ public class ImageFileService : IImageFileService
 
     private readonly string _appDataPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), NameMainFolder, NameImageFolder);
 
-    public IEnumerable<string> SaveImagesToDisk(IEnumerable<string> paths) 
-        => paths.Select(SaveImageToDick);
+    public IEnumerable<string> SaveImagesToDisk(IEnumerable<string>? paths) 
+        => paths?.Select(SaveImageToDick);
 
     public string[]? ShowOpenFileDialog(bool multiselect = true)
     {
@@ -27,8 +27,9 @@ public class ImageFileService : IImageFileService
         return openFileDialog.ShowDialog() == DialogResult.OK ? openFileDialog.FileNames : null;
     }
 
-    public string SaveImageToDick(string path)
+    public string SaveImageToDick(string? path)
     {
+        if (path is null) throw new ServiceException("Передаваемое значение не может быть пустым");
         if (IsFileInAppData(path)) return Path.GetFileName(path);
         Directory.CreateDirectory(_appDataPath);
 
@@ -40,22 +41,25 @@ public class ImageFileService : IImageFileService
         return fileName;
     }
 
-    public string GetFullPath(string fileName)
+    public string GetFullPath(string? fileName)
     {
         var fullPath = Path.Combine(_appDataPath, fileName);
         return System.IO.File.Exists(fullPath) ? fullPath : throw new ServiceException("Не найден файл");
     }
 
-    public void DeleteImageFromDisk(string fileName)
+    public void DeleteImageFromDisk(string? fileName)
     {
+        if (!IsFileInAppData(fileName)) return;
         var fullPath = Path.Combine(_appDataPath, fileName);
 
         if (System.IO.File.Exists(fullPath))
             System.IO.File.Delete(fullPath);
     }
 
-    public bool IsFileInAppData(string filePath)
+    public bool IsFileInAppData(string? filePath)
     {
+        if (filePath is null) return false;
+
         var normalizedFilePath = Path.GetFullPath(filePath).TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
         var normalizedAppDataPath = Path.GetFullPath(_appDataPath).TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
